+++ Tienda_Bruma/src/main/java/com/bruma/domain/Categoria.java
@@ -12,17 +12,17 @@ public class Categoria implements Serializable{
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="id_categoria")
-    private Long idCategoria;
+    @Column(name = "id_categoria")
+    private Integer idCategoria;
     private String descripcion;
     private String rutaImagen;
     private boolean activo;
 
-    public Long getIdCategoria() {
+    public Integer getIdCategoria() {
         return idCategoria;
     }
 
-    public void setIdCategoria(Long idCategoria) {
+    public void setIdCategoria(Integer idCategoria) {
         this.idCategoria = idCategoria;
     }
